Turma model e controller

+++ src/models/inscricaoModel.js
@@ -3,38 +3,30 @@ const db = require('./dbconnection');
 const inscricao = {
 
     /* Função simples que realiza um insert usando o ID do aluno logado e o numero da turma */
-    subscribeClass: async (user_id, class_num, _callback) => {
+    subscribeClass: async (user_id, user_registration, class_num, _callback) => {
 
-        db.query('INSERT INTO inscricoes (id_aluno, numero_turma) VALUES (?, ?)', [user_id, class_num])
-
-    },
-
-    /* Função que faz um select junto de uma sequencia de consultas para retornar um objeto contendo as turmas
-    e suas informaçõesque o aluno logado está inscrito */
-    checkClasses: async (user_id, _callback) => {
-       
-        const query = `
-            SELECT inscricoes.id, turma.numero_turma, turma.turno, disciplina.nome, professor.nome 
-            FROM inscricoes 
-            LEFT JOIN turma 
-            ON inscricoes.numero_turma = turma.numero_turma 
-            INNER JOIN professor 
-            ON turma.id_professor = professor.id 
-            INNER JOIN disciplina 
-            ON turma.id_disciplina = disciplina.id 
-            AND inscricoes.id_aluno = (?)
-        `
-
-        const queryResult = await db.query(query, [user_id])
-        
-        return queryResult[0];
+        const isStudent = await db.query('SELECT * FROM aluno WHERE matricula = (?)', [user_registration]);
+        if (isStudent[0].length > 0) {
+            console.log('Inscrição realizada.')
+            await db.query('INSERT INTO inscricoes (id_aluno, numero_turma) VALUES (?, ?)', [user_id, class_num])
+        } else {
+            console.log('É necessário ser aluno para se inscrever em uma turma')
+        }
 
     },
 
     /* Função que apenas deleta a inscrição do aluno logado da turma selecionada */
-    unsubscribeClass: async (user_id, class_num, _callback) => {
+    unsubscribeClass: async (user_id, user_registration, class_num, _callback) => {
 
-        db.query('DELETE FROM inscricoes WHERE id_aluno = (?) AND numero_turma = (?)', [user_id, class_num])
+        const isStudent = await db.query('SELECT * FROM aluno WHERE matricula = (?)', [user_registration]);
+        console.log(isStudent[0])
+        
+        if (isStudent[0].length > 0) {
+            console.log('Deletado.')
+            await db.query('DELETE FROM inscricoes WHERE id_aluno = (?) AND numero_turma = (?)', [user_id, class_num])
+        } else {
+            console.log('É necessário ser aluno para se desinscrever em uma turma')
+        }
     
     }
